+++ src/main/java/com/kkumteul/domain/book/controller/AdminBookController.java
@@ -3,12 +3,14 @@ package com.kkumteul.domain.book.controller;
 import com.kkumteul.domain.book.dto.AdminInsertBookRequestDto;
 import com.kkumteul.domain.book.service.BookService;
 import com.kkumteul.util.ApiUtil;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
 import static com.kkumteul.util.ApiUtil.ApiSuccess;
 
 @Slf4j
@@ -21,10 +23,9 @@ public class AdminBookController {
 
     // 1. 도서 등록
     @PostMapping
-    public ApiSuccess<?> insertBook(@RequestBody AdminInsertBookRequestDto adminInsertBookRequestDto) {
+    public ApiSuccess<?> insertBook(@Valid @RequestBody AdminInsertBookRequestDto adminInsertBookRequestDto) {
         bookService.insertBook(adminInsertBookRequestDto);
 
         return ApiUtil.success("book insert successfully");
     }
-
 }
