+++ src/main/java/com/kkumteul/exception/GlobalExceptionHandler.java
@@ -1,0 +1,54 @@
+package com.kkumteul.exception.handler;
+
+import com.kkumteul.exception.BookNotFoundException;
+import com.kkumteul.util.ApiUtil;
+import com.kkumteul.util.ApiUtil.ApiError;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.context.request.WebRequest;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@Slf4j
+@RestControllerAdvice
+public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
+
+    @Override
+    protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex,
+                                                                  HttpHeaders headers,
+                                                                  HttpStatusCode status,
+                                                                  WebRequest request) {
+        log.error("Validation failed: " + ex.getMessage(), ex);
+
+        // 유효성 검사 실패 메시지를 수집
+        Map<String, String> errors = new HashMap<>();
+        ex.getBindingResult().getFieldErrors().forEach(error ->
+                errors.put(error.getField(), error.getDefaultMessage())
+        );
+
+        // ApiError 객체를 생성하여 반환
+        ApiError<String> error = ApiUtil.error(HttpServletResponse.SC_BAD_REQUEST, "Validation error");
+
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(error);
+    }
+
+    // 도서 조회 실패 예외
+    @ExceptionHandler({
+        BookNotFoundException.class
+    })
+    protected ResponseEntity<?> handleIllegalArgumentException(Exception e) {
+        log.error(e.getMessage(), e);
+        ApiError<String> error = ApiUtil.error(HttpServletResponse.SC_NOT_FOUND, e.getMessage());
+        return ResponseEntity.status(HttpServletResponse.SC_NOT_FOUND).body(error);
+    }
+
+}
